requirements.txt
amqp==2.5.2
asgiref==3.2.7
beautifulsoup4==4.9.1
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
CouchDB==1.2
Django==3.0.6
django-audiofield==0.10.0
django-celery==3.3.1
django-couchdb-storage==0.0.2
djangorestframework==3.11.0
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
kombu==4.6.8
pyttsx3==2.87
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
